refactor(FileUploaderV2): move file list logic to parent component
